cleaning up functions and changing script
